separate searchbar from searchpage
