+++ client/app/search/search.module.css
@@ -2,41 +2,6 @@
     height: 100vh;
 }
 
-.searchWrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.searchboxWrapper {
-    background-color: rgba(255, 255, 255, 0.171);
-    /* display: flex; */
-    align-items: center;
-    margin-top: 15px;
-    width: 370px;
-    height: 40px;
-    border-radius: 5px;
-}
-
-.searchbox {
-    border: none;
-    width: 340px;
-    height: 40px;
-    color: white;
-    font-size: 18px;
-    background-color: rgba(255, 255, 255, 0);
-}
-
-.searchbox::placeholder {
-    color: rgb(161, 161, 161);
-    font-size: 18px;
-  }
-
-.magnifyingglass{
-    width: 20px;
-    margin-right: 10px;
-}
-
 .title {
     margin-left: 20px;
     margin-top: 24px;
@@ -50,10 +15,6 @@
     margin-top: 30px;
     font-size: 14px;
     font-weight: bold;
-}
-
-.submitbtn {
-    all: unset;
 }
 
 .num {
