+++ client/components/Search/PopularKeyword.ts
@@ -1,13 +1,16 @@
 // 인기 검색어 get 함수
 
 const getSearchData = async () => {
-    const res = await fetch('http://localhost:3000/api/popularkeyword');
 
-    if (!res.ok) {
-        throw new Error('failed to data fetch')
+    try {
+        const response = await fetch(`http://localhost:3000/api/popularkeyword`);
+        const { data } = await response.json();
+        return data;
+
+    } catch (error) {
+        console.error(error);
+        return [];
     }
-
-    return await res.json();
 }
 
 export default getSearchData;
