+++ client/public/fallback-SvZZugCeYTn2kDTQtIrpg.js
@@ -1,0 +1,1 @@
+(()=>{"use strict";self.fallback=async e=>"document"===e.destination?caches.match("/_offline",{ignoreSearch:!0}):Response.error()})();
